Handled deprecated ES stats (#61)
* fixed logic for removing deprecated metrics

* string to int for version numbers

* added unit test

* add dependency

* added dependency

* added dependency

* fixed test errors

* mocked module collectd

* fixed docker build error

* edited README file

* halting execution on unit test failure

* halting execution on unit test failure

* halting execution on unit test failure

* halting execution on unit test failure

* halting execution on unit test failure

* halting execution on unit test failure

* halting execution on unit test failure

* halting execution on unit test failure

+++ tests/run_tests.sh
@@ -8,6 +8,13 @@ PYTHON_VERSION=$(${PYTHON} -V 2>&1)
 
 echo "Interpreter version: ${PYTHON_VERSION}"
 
+echo "Running unit tests"
+pytest ../elasticsearch_collectd_test.py
+status=$?
+if [[ "$status" != 0 ]]; then
+    exit ${status}
+fi
+
 tmpfile=$(mktemp /tmp/run_tests.sh.XXXXXX)
 trap 'rm -f $tmpfile' 1 2 3 15
 for scenario in `ls data`; do
